chore: Move Primitives to component directory
